[pdata] Move pack_node to btree
